feat: add ImageWithFallback
