+++ apps/web/src/features/products/components/ProductTrendingImageCard.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ProductTrendingImageCard = () => {
+  return <div></div>;
+};
+
+export default ProductTrendingImageCard;
